Add a test for the doFetchIp saga
I'm not entirely convinced that I like this way of testing.
Returning declarative javascript objects is a nice property, but the nature
of iterators means you need to have internal knowledge of the sequence of steps.
So we're not only testing the contract of the saga; we're testing the internals.